added support to run single testcase of needed
